+++ src/main/java/br/com/fiap/oceanGuardian/repository/EventoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.fiap.oceanGuardian.model.Evento;
 
-public interface EventoRepository extends JpaRepository<Evento, Long>{
+public interface EventoRepository extends JpaRepository<Evento, Integer>{
     
 }
