+++ src/main/java/br/com/fiap/oceanGuardian/repository/ReporteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.fiap.oceanGuardian.model.Reporte;
 
-public interface ReporteRepository extends JpaRepository<Reporte, Long> {
+public interface ReporteRepository extends JpaRepository<Reporte, Integer> {
     
 }
